+++ src/com/landawn/abacus/util/ContinuableFuture.java
@@ -209,19 +209,19 @@ public class ContinuableFuture<T> implements Future<T> {
         return future.get(timeout, unit);
     }
 
-    public Pair<T, Exception> gett() {
+    public Result<T, Exception> gett() {
         try {
-            return Pair.of(get(), null);
+            return Result.of(get(), null);
         } catch (Exception e) {
-            return Pair.of(null, e);
+            return Result.of(null, e);
         }
     }
 
-    public Pair<T, Exception> gett(final long timeout, final TimeUnit unit) {
+    public Result<T, Exception> gett(final long timeout, final TimeUnit unit) {
         try {
-            return Pair.of(get(timeout, unit), null);
+            return Result.of(get(timeout, unit), null);
         } catch (Exception e) {
-            return Pair.of(null, e);
+            return Result.of(null, e);
         }
     }
 
@@ -239,14 +239,14 @@ public class ContinuableFuture<T> implements Future<T> {
     }
 
     public <U, E extends Exception> U getThenApply(final Try.BiFunction<? super T, ? super Exception, ? extends U, E> action) throws E {
-        final Pair<T, Exception> result = gett();
-        return action.apply(result.left, result.right);
+        final Result<T, Exception> result = gett();
+        return action.apply(result.orElse(null), result.getExceptionIfPresent());
     }
 
     public <U, E extends Exception> U getThenApply(final long timeout, final TimeUnit unit,
             final Try.BiFunction<? super T, ? super Exception, ? extends U, E> action) throws E {
-        final Pair<T, Exception> result = gett(timeout, unit);
-        return action.apply(result.left, result.right);
+        final Result<T, Exception> result = gett(timeout, unit);
+        return action.apply(result.orElse(null), result.getExceptionIfPresent());
     }
 
     public <E extends Exception> void getThenAccept(final Try.Consumer<? super T, E> action) throws InterruptedException, ExecutionException, E {
@@ -259,14 +259,14 @@ public class ContinuableFuture<T> implements Future<T> {
     }
 
     public <E extends Exception> void getThenAccept(final Try.BiConsumer<? super T, ? super Exception, E> action) throws E {
-        final Pair<T, Exception> result = gett();
-        action.accept(result.left, result.right);
+        final Result<T, Exception> result = gett();
+        action.accept(result.orElse(null), result.getExceptionIfPresent());
     }
 
     public <E extends Exception> void getThenAccept(final long timeout, final TimeUnit unit, final Try.BiConsumer<? super T, ? super Exception, E> action)
             throws E {
-        final Pair<T, Exception> result = gett(timeout, unit);
-        action.accept(result.left, result.right);
+        final Result<T, Exception> result = gett(timeout, unit);
+        action.accept(result.orElse(null), result.getExceptionIfPresent());
     }
 
     //    public void complete() throws InterruptedException, ExecutionException {
@@ -282,8 +282,8 @@ public class ContinuableFuture<T> implements Future<T> {
     //    }
     //
     //    public void complete(BiConsumer<? super T, ? super Exception> action) {
-    //        final Pair<T, Exception> result = gett();
-    //        action.accept(result.left, result.right);
+    //        final Result<T, Exception> result = gett();
+    //        action.accept(result.orElse(null), result.getExceptionIfPresent());
     //    }
 
     <U, E extends Exception> ContinuableFuture<U> thenApply(final Try.Function<? super T, ? extends U, E> action) {
@@ -356,14 +356,14 @@ public class ContinuableFuture<T> implements Future<T> {
     //
     //            @Override
     //            public U get() throws InterruptedException, ExecutionException {
-    //                final Pair<T, Exception> result = gett();
-    //                return action.apply(result.left, result.right);
+    //                final Result<T, Exception> result = gett();
+    //                return action.apply(result.orElse(null), result.getExceptionIfPresent());
     //            }
     //
     //            @Override
     //            public U get(final long timeout, final TimeUnit unit) throws InterruptedException, ExecutionException, TimeoutException {
-    //                final Pair<T, Exception> result = gett(timeout, unit);
-    //                return action.apply(result.left, result.right);
+    //                final Result<T, Exception> result = gett(timeout, unit);
+    //                return action.apply(result.orElse(null), result.getExceptionIfPresent());
     //            }
     //        }, null, asyncExecutor) {
     //            @Override
@@ -463,10 +463,10 @@ public class ContinuableFuture<T> implements Future<T> {
     //
     //            @Override
     //            public R get() throws InterruptedException, ExecutionException {
-    //                final Pair<T, Exception> result = gett();
+    //                final Result<T, Exception> result = gett();
     //                final Pair<? extends U, ? super Exception> result2 = other.gett();
     //
-    //                return action.apply(Tuple.of(result.left, result.right, (U) result2.left, result2.right));
+    //                return action.apply(Tuple.of(result.orElse(null), result.getExceptionIfPresent(), (U) result2.orElse(null), result2.getExceptionIfPresent()));
     //            }
     //
     //            @Override
@@ -475,10 +475,10 @@ public class ContinuableFuture<T> implements Future<T> {
     //                final long now = N.currentMillis();
     //                final long endTime = timeoutInMillis > Long.MAX_VALUE - now ? Long.MAX_VALUE : now + timeoutInMillis;
     //
-    //                final Pair<T, Exception> result = ContinuableFuture.this.gett(timeout, unit);
+    //                final Result<T, Exception> result = ContinuableFuture.this.gett(timeout, unit);
     //                final Pair<? extends U, ? super Exception> result2 = other.gett(N.max(0, endTime - N.currentMillis()), TimeUnit.MILLISECONDS);
     //
-    //                return action.apply(Tuple.of(result.left, result.right, (U) result2.left, result2.right));
+    //                return action.apply(Tuple.of(result.orElse(null), result.getExceptionIfPresent(), (U) result2.orElse(null), result2.getExceptionIfPresent()));
     //            }
     //        }, null, asyncExecutor) {
     //            @Override
@@ -538,8 +538,8 @@ public class ContinuableFuture<T> implements Future<T> {
         return execute(new Callable<Void>() {
             @Override
             public Void call() throws Exception {
-                final Pair<T, Exception> result = gett();
-                action.accept(result.left, result.right);
+                final Result<T, Exception> result = gett();
+                action.accept(result.orElse(null), result.getExceptionIfPresent());
                 return null;
             }
         });
@@ -568,8 +568,8 @@ public class ContinuableFuture<T> implements Future<T> {
         return execute(new Callable<R>() {
             @Override
             public R call() throws Exception {
-                final Pair<T, Exception> result = gett();
-                return action.apply(result.left, result.right);
+                final Result<T, Exception> result = gett();
+                return action.apply(result.orElse(null), result.getExceptionIfPresent());
             }
         });
     }
@@ -601,10 +601,10 @@ public class ContinuableFuture<T> implements Future<T> {
         return execute(new Callable<Void>() {
             @Override
             public Void call() throws Exception {
-                final Pair<T, Exception> result = gett();
-                final Pair<U, ? super Exception> result2 = other.gett();
+                final Result<T, Exception> result = gett();
+                final Result<U, Exception> result2 = other.gett();
 
-                action.accept(Tuple.of(result.left, result.right, result2.left, result2.right));
+                action.accept(Tuple.of(result.orElse(null), result.getExceptionIfPresent(), result2.orElse(null), result2.getExceptionIfPresent()));
                 return null;
             }
         }, other);
@@ -635,10 +635,10 @@ public class ContinuableFuture<T> implements Future<T> {
         return execute(new Callable<R>() {
             @Override
             public R call() throws Exception {
-                final Pair<T, Exception> result = gett();
-                final Pair<U, ? super Exception> result2 = other.gett();
+                final Result<T, Exception> result = gett();
+                final Result<U, Exception> result2 = other.gett();
 
-                return action.apply(Tuple.of(result.left, result.right, result2.left, result2.right));
+                return action.apply(Tuple.of(result.orElse(null), result.getExceptionIfPresent(), result2.orElse(null), result2.getExceptionIfPresent()));
             }
         }, other);
     }
@@ -672,9 +672,9 @@ public class ContinuableFuture<T> implements Future<T> {
         return execute(new Callable<Void>() {
             @Override
             public Void call() throws Exception {
-                final Pair<T, Exception> result = Futures.anyOf(N.asList(ContinuableFuture.this, other)).gett();
+                final Result<T, Exception> result = Futures.anyOf(N.asList(ContinuableFuture.this, other)).gett();
 
-                action.accept(result.left, result.right);
+                action.accept(result.orElse(null), result.getExceptionIfPresent());
                 return null;
             }
         }, other);
@@ -708,9 +708,9 @@ public class ContinuableFuture<T> implements Future<T> {
         return execute(new Callable<R>() {
             @Override
             public R call() throws Exception {
-                final Pair<T, Exception> result = Futures.anyOf(N.asList(ContinuableFuture.this, other)).gett();
+                final Result<T, Exception> result = Futures.anyOf(N.asList(ContinuableFuture.this, other)).gett();
 
-                return action.apply(result.left, result.right);
+                return action.apply(result.orElse(null), result.getExceptionIfPresent());
             }
         }, other);
     }
@@ -819,11 +819,11 @@ public class ContinuableFuture<T> implements Future<T> {
     //
     //            @Override
     //            public T get() throws InterruptedException, ExecutionException {
-    //                final Pair<T, Exception> result = gett();
+    //                final Result<T, Exception> result = gett();
     //
     //                if (result.right != null) {
     //                    try {
-    //                        action.accept(result.left, result.right);
+    //                        action.accept(result.orElse(null), result.getExceptionIfPresent());
     //                    } catch (Exception e) {
     //                        // ignore.
     //                    }
@@ -836,18 +836,18 @@ public class ContinuableFuture<T> implements Future<T> {
     //                        throw N.toRuntimeException(result.right);
     //                    }
     //                } else {
-    //                    action.accept(result.left, result.right);
+    //                    action.accept(result.orElse(null), result.getExceptionIfPresent());
     //                    return result.left;
     //                }
     //            }
     //
     //            @Override
     //            public T get(final long timeout, final TimeUnit unit) throws InterruptedException, ExecutionException, TimeoutException {
-    //                final Pair<T, Exception> result = gett(timeout, unit);
+    //                final Result<T, Exception> result = gett(timeout, unit);
     //
     //                if (result.right != null) {
     //                    try {
-    //                        action.accept(result.left, result.right);
+    //                        action.accept(result.orElse(null), result.getExceptionIfPresent());
     //                    } catch (Exception e) {
     //                        // ignore.
     //                    }
@@ -860,7 +860,7 @@ public class ContinuableFuture<T> implements Future<T> {
     //                        throw N.toRuntimeException(result.right);
     //                    }
     //                } else {
-    //                    action.accept(result.left, result.right);
+    //                    action.accept(result.orElse(null), result.getExceptionIfPresent());
     //                    return result.left;
     //                }
     //            }
@@ -886,14 +886,14 @@ public class ContinuableFuture<T> implements Future<T> {
     //
     //            @Override
     //            public U get() throws InterruptedException, ExecutionException {
-    //                final Pair<T, Exception> result = gett();
-    //                return action.apply(result.left, result.right);
+    //                final Result<T, Exception> result = gett();
+    //                return action.apply(result.orElse(null), result.getExceptionIfPresent());
     //            }
     //
     //            @Override
     //            public U get(final long timeout, final TimeUnit unit) throws InterruptedException, ExecutionException, TimeoutException {
-    //                final Pair<T, Exception> result = gett(timeout, unit);
-    //                return action.apply(result.left, result.right);
+    //                final Result<T, Exception> result = gett(timeout, unit);
+    //                return action.apply(result.orElse(null), result.getExceptionIfPresent());
     //            }
     //        }, asyncExecutor);
     //    }